Allocator: fix integer promotion warnings in FSlabHeap

+++ Source/Runtime/Core/Public/Allocator/SlabHeap.h
@@ -208,7 +208,7 @@ public:
     NODISCARD PPE_DECLSPEC_ALLOCATOR() TMemoryView<T> AllocateCopyT(TMemoryView<T> src) {
         TMemoryView<Meta::TRemoveConst<T>> dst;
         if (not src.empty()) {
-            dst = AllocateT<Meta::TRemoveConst<T>>(src.size());
+            dst = AllocateT<Meta::TRemoveConst<T>>(checked_cast<u32>(src.size()));
             src.CopyTo(dst);
         }
         return dst;
@@ -219,7 +219,7 @@ public:
         using map_type = decltype(std::declval<_Map>()(std::declval<T&>()));
         TMemoryView<map_type> dst;
         if (not src.empty()) {
-            dst = AllocateT<map_type>(src.size());
+            dst = AllocateT<map_type>(checked_cast<u32>(src.size()));
             src.Map(std::forward<_Map>(map)).UnitializedCopyTo(dst.begin());
         }
         return dst;
